Added TileInteractionActivity to which will be a base for all activities that need to move to a tile and interact with it.
