+++ core/src/com/itemshop/factories/characters/ShopkeeperFactory.java
@@ -1,7 +1,6 @@
 package com.itemshop.factories.characters;
 
 import java.util.Random;
-
 import com.badlogic.ashley.core.Engine;
 import com.badlogic.ashley.core.Entity;
 import com.itemshop.character.Character;
