+++ core/src/com/itemshop/schedule/activities/AddItemToContainerActivity.java
@@ -3,15 +3,13 @@ package com.itemshop.schedule.activities;
 import com.badlogic.ashley.core.ComponentMapper;
 import com.badlogic.ashley.core.Entity;
 import com.itemshop.container.ContainerComponent;
-import com.itemshop.render.PositionComponent;
 
 /**
  * An activity which involves walking to a container an placing the specified item into it.
  */
-public class AddItemToContainerActivity extends WalkActivity {
+public class AddItemToContainerActivity extends TileInteractionActivity {
 	
 	/** The required component mappers. */
-    private static ComponentMapper<PositionComponent> positionMapper   = ComponentMapper.getFor(PositionComponent.class);
     private static ComponentMapper<ContainerComponent> containerMapper = ComponentMapper.getFor(ContainerComponent.class);
     
     /** The target container. */
@@ -27,49 +25,21 @@ public class AddItemToContainerActivity extends WalkActivity {
      * @param item
      */
 	public AddItemToContainerActivity(Entity character, Entity container, Entity item) {
-		// Call through to the walk activity to set our target location to match that of the target container.
-		super(character, (int) positionMapper.get(container).x, (int) positionMapper.get(container).y);
+		super(character, container);
 		this.container = container;
 		this.item      = item;
 	}
 	
 	/**
-	 * Add the item to the container.
+	 * Called when the tile to interact with has been reached.
 	 */
-	public void addItemToContainer() {
+	@Override
+	public void interact() {
+		// Add the item to the container.
 		try {
 			containerMapper.get(container).add(item);
 		} catch (Exception e) {
-			// If an exception was thrown it was because there is no room for this item. Do nothing.
+			// If an exception was thrown it was because there is no room for this item. Do nothing for now.
 		}
 	}
-
-	@Override
-	public void onBegin() {
-		// Begin the walking activity.
-		super.onBegin();
-	}
-
-	@Override
-	public void perform() {
-		// We need to wait for a path to the target.
-		if (this.getPath().isPathComputed) {
-			// If the path is blocked then we cannot reach our container and ...
-			if (this.getPath().isPathBlocked) {
-				// ... we cannot do anything.
-				this.finish();
-			} else if (this.getPath().isComplete) {
-				// We have reached our target, place the item ... 
-				addItemToContainer();
-				// ... and then we are done.
-				this.finish();
-			}
-		}
-	}
-
-	@Override
-	public void onInterrupt() {}
-
-	@Override
-	public void onEnd() {}
 }
